+++ src/routes/datasets/Datasets.tsx
@@ -14,11 +14,7 @@ import useDispatchWithAutoAbort from '@/hooks/useDispatchWithAutoAbort';
 import { useDocumentTitleEffect } from '@/hooks/useDocumentTitleEffect';
 import useLocationWithParams from '@/hooks/useLocationWithParams';
 import useSearchFiltersEffect from '@/hooks/useSearchFiltersEffect';
-import {
-    useSyncedDateStringState,
-    useSyncedStringArrayState,
-    useSyncedStringState,
-} from '@/hooks/useSyncedState';
+import { useCreationDate, useOrdering, useOwner } from '@/hooks/useSyncedState';
 import {
     TableFiltersContext,
     useTableFiltersContext,
@@ -55,16 +51,9 @@ const Datasets = (): JSX.Element => {
         params: { page, search: searchFilters, match },
         setLocationWithParams,
     } = useLocationWithParams();
-    const [ordering] = useSyncedStringState('ordering', '-creation_date');
-    const [owner] = useSyncedStringArrayState('owner', []);
-    const [creation_date_after] = useSyncedDateStringState(
-        'creation_date_after',
-        ''
-    );
-    const [creation_date_before] = useSyncedDateStringState(
-        'creation_date_before',
-        ''
-    );
+    const [ordering] = useOrdering('-creation_date');
+    const [owner] = useOwner();
+    const { creationDateAfter, creationDateBefore } = useCreationDate();
 
     useSearchFiltersEffect(() => {
         return dispatchWithAutoAbort(
@@ -74,8 +63,8 @@ const Datasets = (): JSX.Element => {
                 ordering,
                 match,
                 owner__in: owner,
-                creation_date_after,
-                creation_date_before,
+                creationDateAfter,
+                creationDateBefore,
             })
         );
     }, [
@@ -84,8 +73,8 @@ const Datasets = (): JSX.Element => {
         ordering,
         match,
         owner,
-        creation_date_after,
-        creation_date_before,
+        creationDateAfter,
+        creationDateBefore,
     ]);
 
     const datasets: DatasetStubType[] = useAppSelector(