+++ src/routes/tasks/Tasks.tsx
@@ -7,9 +7,12 @@ import { useAssetListDocumentTitleEffect } from '@/hooks/useDocumentTitleEffect'
 import useKeyFromPath from '@/hooks/useKeyFromPath';
 import useLocationWithParams from '@/hooks/useLocationWithParams';
 import {
-    useSyncedDateStringState,
-    useSyncedStringArrayState,
-    useSyncedStringState,
+    useCreationDate,
+    useEndDate,
+    useOrdering,
+    useStartDate,
+    useStatus,
+    useWorker,
 } from '@/hooks/useSyncedState';
 import {
     listAggregateTasks,
@@ -75,24 +78,12 @@ const TestTasks = ({ taskKey }: TasksProps): JSX.Element => {
         params: { search: searchFilters, page, match },
         setLocationWithParams,
     } = useLocationWithParams();
-    const [ordering] = useSyncedStringState('ordering', '-rank');
-    const [status] = useSyncedStringArrayState('status', []);
-    const [worker] = useSyncedStringArrayState('worker', []);
-    const [creation_date_after] = useSyncedDateStringState(
-        'creation_date_after',
-        ''
-    );
-    const [creation_date_before] = useSyncedDateStringState(
-        'creation_date_before',
-        ''
-    );
-    const [start_date_after] = useSyncedDateStringState('start_date_after', '');
-    const [start_date_before] = useSyncedDateStringState(
-        'start_date_before',
-        ''
-    );
-    const [end_date_after] = useSyncedDateStringState('end_date_after', '');
-    const [end_date_before] = useSyncedDateStringState('end_date_before', '');
+    const [ordering] = useOrdering('-rank');
+    const [status] = useStatus();
+    const [worker] = useWorker();
+    const { creationDateBefore, creationDateAfter } = useCreationDate();
+    const { startDateBefore, startDateAfter } = useStartDate();
+    const { endDateBefore, endDateAfter } = useEndDate();
 
     const loading = useAppSelector((state) => state.tasks.testTasksLoading);
     const list = () =>
@@ -103,12 +94,12 @@ const TestTasks = ({ taskKey }: TasksProps): JSX.Element => {
             match,
             status,
             worker__in: worker,
-            creation_date_after,
-            creation_date_before,
-            start_date_after,
-            start_date_before,
-            end_date_after,
-            end_date_before,
+            creationDateAfter,
+            creationDateBefore,
+            startDateAfter,
+            startDateBefore,
+            endDateAfter,
+            endDateBefore,
         });
     const tasks = useAppSelector((state) => state.tasks.testTasks);
     const count = useAppSelector((state) => state.tasks.testTasksCount);
@@ -147,24 +138,12 @@ const TrainTasks = ({ taskKey }: TasksProps): JSX.Element => {
         params: { search: searchFilters, page, match },
         setLocationWithParams,
     } = useLocationWithParams();
-    const [ordering] = useSyncedStringState('ordering', '-rank');
-    const [status] = useSyncedStringArrayState('status', []);
-    const [worker] = useSyncedStringArrayState('worker', []);
-    const [creation_date_after] = useSyncedDateStringState(
-        'creation_date_after',
-        ''
-    );
-    const [creation_date_before] = useSyncedDateStringState(
-        'creation_date_before',
-        ''
-    );
-    const [start_date_after] = useSyncedDateStringState('start_date_after', '');
-    const [start_date_before] = useSyncedDateStringState(
-        'start_date_before',
-        ''
-    );
-    const [end_date_after] = useSyncedDateStringState('end_date_after', '');
-    const [end_date_before] = useSyncedDateStringState('end_date_before', '');
+    const [ordering] = useOrdering('-rank');
+    const [status] = useStatus();
+    const [worker] = useWorker();
+    const { creationDateBefore, creationDateAfter } = useCreationDate();
+    const { startDateBefore, startDateAfter } = useStartDate();
+    const { endDateBefore, endDateAfter } = useEndDate();
 
     const loading = useAppSelector((state) => state.tasks.trainTasksLoading);
     const list = () =>
@@ -175,12 +154,12 @@ const TrainTasks = ({ taskKey }: TasksProps): JSX.Element => {
             match,
             status,
             worker__in: worker,
-            creation_date_after,
-            creation_date_before,
-            start_date_after,
-            start_date_before,
-            end_date_after,
-            end_date_before,
+            creationDateAfter,
+            creationDateBefore,
+            startDateAfter,
+            startDateBefore,
+            endDateAfter,
+            endDateBefore,
         });
     const tasks = useAppSelector((state) => state.tasks.trainTasks);
     const count = useAppSelector((state) => state.tasks.trainTasksCount);
@@ -219,24 +198,12 @@ const CompositeTrainTasks = ({ taskKey }: TasksProps): JSX.Element => {
         params: { search: searchFilters, page, match },
         setLocationWithParams,
     } = useLocationWithParams();
-    const [ordering] = useSyncedStringState('ordering', '-rank');
-    const [status] = useSyncedStringArrayState('status', []);
-    const [worker] = useSyncedStringArrayState('worker', []);
-    const [creation_date_after] = useSyncedDateStringState(
-        'creation_date_after',
-        ''
-    );
-    const [creation_date_before] = useSyncedDateStringState(
-        'creation_date_before',
-        ''
-    );
-    const [start_date_after] = useSyncedDateStringState('start_date_after', '');
-    const [start_date_before] = useSyncedDateStringState(
-        'start_date_before',
-        ''
-    );
-    const [end_date_after] = useSyncedDateStringState('end_date_after', '');
-    const [end_date_before] = useSyncedDateStringState('end_date_before', '');
+    const [ordering] = useOrdering('-rank');
+    const [status] = useStatus();
+    const [worker] = useWorker();
+    const { creationDateBefore, creationDateAfter } = useCreationDate();
+    const { startDateBefore, startDateAfter } = useStartDate();
+    const { endDateBefore, endDateAfter } = useEndDate();
 
     const loading = useAppSelector(
         (state) => state.tasks.compositeTasksLoading
@@ -249,12 +216,12 @@ const CompositeTrainTasks = ({ taskKey }: TasksProps): JSX.Element => {
             match,
             status,
             worker__in: worker,
-            creation_date_after,
-            creation_date_before,
-            start_date_after,
-            start_date_before,
-            end_date_after,
-            end_date_before,
+            creationDateAfter,
+            creationDateBefore,
+            startDateAfter,
+            startDateBefore,
+            endDateAfter,
+            endDateBefore,
         });
     const tasks = useAppSelector((state) => state.tasks.compositeTasks);
     const count = useAppSelector((state) => state.tasks.compositeTasksCount);
@@ -293,24 +260,12 @@ const AggregateTasks = ({ taskKey }: TasksProps): JSX.Element => {
         params: { search: searchFilters, page, match },
         setLocationWithParams,
     } = useLocationWithParams();
-    const [ordering] = useSyncedStringState('ordering', '-rank');
-    const [status] = useSyncedStringArrayState('status', []);
-    const [worker] = useSyncedStringArrayState('worker', []);
-    const [creation_date_after] = useSyncedDateStringState(
-        'creation_date_after',
-        ''
-    );
-    const [creation_date_before] = useSyncedDateStringState(
-        'creation_date_before',
-        ''
-    );
-    const [start_date_after] = useSyncedDateStringState('start_date_after', '');
-    const [start_date_before] = useSyncedDateStringState(
-        'start_date_before',
-        ''
-    );
-    const [end_date_after] = useSyncedDateStringState('end_date_after', '');
-    const [end_date_before] = useSyncedDateStringState('end_date_before', '');
+    const [ordering] = useOrdering('-rank');
+    const [status] = useStatus();
+    const [worker] = useWorker();
+    const { creationDateBefore, creationDateAfter } = useCreationDate();
+    const { startDateBefore, startDateAfter } = useStartDate();
+    const { endDateBefore, endDateAfter } = useEndDate();
 
     const loading = useAppSelector(
         (state) => state.tasks.aggregateTasksLoading
@@ -323,12 +278,12 @@ const AggregateTasks = ({ taskKey }: TasksProps): JSX.Element => {
             match,
             status,
             worker__in: worker,
-            creation_date_after,
-            creation_date_before,
-            start_date_after,
-            start_date_before,
-            end_date_after,
-            end_date_before,
+            creationDateAfter,
+            creationDateBefore,
+            startDateAfter,
+            startDateBefore,
+            endDateAfter,
+            endDateBefore,
         });
     const tasks = useAppSelector((state) => state.tasks.aggregateTasks);
     const count = useAppSelector((state) => state.tasks.aggregateTasksCount);